refactor(download): integrate @kesha-antonov/react-native-background-downloader for background model downloads and enhance documentation
