+++ src/download/constants.ts
@@ -1,0 +1,5 @@
+export const RELEASE_API_BASE =
+  'https://api.github.com/repos/k2-fsa/sherpa-onnx/releases/tags';
+export const CACHE_TTL_MINUTES = 24 * 60;
+export const MODEL_ARCHIVE_EXT = '.tar.bz2';
+export const MODEL_ONNX_EXT = '.onnx';
